fix: show sponsored fees in Swap history details (OK-58376)
