+++ packages/kit/src/views/Swap/utils/swapGasUtils.test.ts
@@ -1,0 +1,56 @@
+import {
+  isSwapGasAccountSponsored,
+  isSwapGasSponsored,
+  isSwapMegafuelSponsored,
+} from './swapGasUtils';
+
+describe('isSwapGasSponsored', () => {
+  it('recognizes final estimate-fee sponsorship signals', () => {
+    expect(
+      isSwapGasSponsored({
+        gasAccountEligible: true,
+        payer: 'gasAccount',
+        gasAccountQuote: {
+          quoteId: 'quote-id',
+          maxFee: '1',
+          expiresAt: String(Date.now() + 60_000),
+        },
+      }),
+    ).toBe(true);
+    expect(
+      isSwapGasSponsored({
+        megafuelEligible: { sponsorable: true, sponsorName: 'OneKey' },
+      }),
+    ).toBe(true);
+    expect(isSwapGasSponsored({ payer: 'megafuel' })).toBe(true);
+  });
+
+  it('does not mark a user-paid fee as sponsored', () => {
+    expect(isSwapGasSponsored({ gasAccountEligible: true })).toBe(false);
+    expect(
+      isSwapGasSponsored({
+        gasAccountEligible: true,
+        payer: 'gasAccount',
+      }),
+    ).toBe(false);
+    expect(isSwapGasSponsored({ payer: 'user' })).toBe(false);
+    expect(isSwapGasSponsored()).toBe(false);
+  });
+
+  it('separates MegaFuel sponsorship from Gas Account sponsorship', () => {
+    expect(isSwapMegafuelSponsored({ gasAccountEligible: true })).toBe(false);
+    expect(isSwapMegafuelSponsored({ payer: 'megafuel' })).toBe(true);
+    expect(isSwapGasAccountSponsored({ gasAccountEligible: true })).toBe(false);
+    expect(
+      isSwapGasAccountSponsored({
+        gasAccountEligible: true,
+        payer: 'gasAccount',
+        gasAccountQuote: {
+          quoteId: 'quote-id',
+          maxFee: '1',
+          expiresAt: String(Date.now() + 60_000),
+        },
+      }),
+    ).toBe(true);
+  });
+});
